LayoutTests/imported/w3c/web-platform-tests/navigation-api/navigation-methods/return-value/navigate-pagehide-expected.txt

CONSOLE MESSAGE: Unhandled Promise Rejection: Error: assert_throws_dom: function "() => { throw committedReason; }" threw object "AbortError: Navigation aborted" that is not a DOMException InvalidStateError: property "code" is equal to 20, expect...


Harness Error (FAIL), message = Unhandled rejection: assert_throws_dom: function "() => { throw committedReason; }" threw object "AbortError: Navigation aborted" that is not a DOMException InvalidStateError: property "code" is equal to 20, expected 11

TIMEOUT navigate() inside onpagehide Test timed out

Harness Error (FAIL), message = Unhandled rejection: assert_throws_dom: function "() => { throw committedReason; }" threw object "AbortError: Navigation aborted" that is not a DOMException InvalidStateError: property "code" is equal to 20, expected 11

TIMEOUT navigate() inside onpagehide Test timed out
